D2 relative strategy refactor files (#311)
* [Relative Load Balancing] refactor common files: ring related factories and quaratine
