+++ d2/src/main/java/com/linkedin/d2/balancer/config/RelativeStrategyPropertiesConverter.java
@@ -26,7 +26,7 @@ import com.linkedin.d2.HttpMethod;
 import com.linkedin.d2.HttpStatusCodeRange;
 import com.linkedin.d2.HttpStatusCodeRangeArray;
 import com.linkedin.d2.balancer.properties.PropertyKeys;
-import com.linkedin.d2.balancer.strategies.degrader.DelegatingRingFactory;
+import com.linkedin.d2.balancer.strategies.DelegatingRingFactory;
 import com.linkedin.d2.balancer.util.hashing.URIRegexHash;
 import com.linkedin.data.codec.JacksonDataCodec;
 import com.linkedin.data.schema.validation.CoercionMode;
