+++ d2/src/main/java/com/linkedin/d2/balancer/strategies/DistributionNonDiscreteRingFactory.java
@@ -1,0 +1,29 @@
+/*
+   Copyright (c) 2018 LinkedIn Corp.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package com.linkedin.d2.balancer.strategies;
+
+import com.linkedin.d2.balancer.util.hashing.DistributionNonDiscreteRing;
+import com.linkedin.d2.balancer.util.hashing.Ring;
+import java.util.Map;
+
+
+public class DistributionNonDiscreteRingFactory<T> implements RingFactory<T> {
+  @Override
+  public Ring<T> createRing(Map<T, Integer> points) {
+    return new DistributionNonDiscreteRing<>(points);
+  }
+}
