+++ d2/src/main/java/com/linkedin/d2/balancer/strategies/PointBasedConsistentHashRingFactory.java
@@ -14,8 +14,9 @@
    limitations under the License.
 */
 
-package com.linkedin.d2.balancer.strategies.degrader;
+package com.linkedin.d2.balancer.strategies;
 
+import com.linkedin.d2.balancer.strategies.degrader.DegraderLoadBalancerStrategyConfig;
 import com.linkedin.d2.balancer.util.hashing.ConsistentHashRing;
 import com.linkedin.d2.balancer.util.hashing.ConsistentHashRing.Point;
 import com.linkedin.d2.balancer.util.hashing.Ring;
