+++ d2/src/main/java/com/linkedin/d2/balancer/util/hashing/BoundedLoadConsistentHashRing.java
@@ -16,7 +16,7 @@
 
 package com.linkedin.d2.balancer.util.hashing;
 
-import com.linkedin.d2.balancer.strategies.degrader.RingFactory;
+import com.linkedin.d2.balancer.strategies.RingFactory;
 import com.linkedin.util.degrader.CallTracker;
 import java.util.ArrayList;
 import java.util.Collections;
