+++ d2/src/test/java/com/linkedin/d2/balancer/strategies/DelegatingRingFactoryTest.java
@@ -14,9 +14,10 @@
    limitations under the License.
 */
 
-package com.linkedin.d2.balancer.strategies.degrader;
+package com.linkedin.d2.balancer.strategies;
 
 
+import com.linkedin.d2.balancer.strategies.degrader.DegraderLoadBalancerStrategyConfig;
 import com.linkedin.d2.balancer.util.hashing.ConsistentHashRing;
 import com.linkedin.d2.balancer.util.hashing.ConsistentHashRing.Point;
 import com.linkedin.d2.balancer.util.hashing.DistributionNonDiscreteRing;
