+++ d2/src/test/java/com/linkedin/d2/balancer/util/hashing/ConsistentHashRingTest.java
@@ -21,8 +21,8 @@ import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
 
+import com.linkedin.d2.balancer.strategies.DelegatingRingFactory;
 import com.linkedin.d2.balancer.strategies.degrader.DegraderLoadBalancerStrategyConfig;
-import com.linkedin.d2.balancer.strategies.degrader.DelegatingRingFactory;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.HashMap;
